requirements.txt
torch>=2.1.0
torchvision
transformers>=4.46.0
Pillow>=10.0.0
tqdm
h5py
